Attempt to use grid layout for #12
I'm aware the prefferred way to add content is to use semantinc html
with nav, header, main etc.

I tried to use it with grid and it was mostly broken, I still need
to learn how to do it properly.

In the meantime I added this using just the top level elements, it would
be nice to see how can this be changed to use semantic html.