Very basic (and dangerous) save to file.
This is just to demostrate the following:

Use of /lib folder, here is where the  files containing logic for certain
operations will live. In this example there's an "entries.js" file that
has the logic to create a new entry (and in the future, update it,
delete it etc)  This is better than just adding all the logic into the
/controller/api.js file, whose soley responsability should be take data
from the browser and ask some other library to do something with it

Use of a /data folder (well in this case /entries). This is where data
generated by the app will be stored, usually this is not persisted in
the git repo (it could expand to millons of records). To keep things
simple I'm adding it here

entries/index.json This is our master data source, it contains what our
blogs should display, every time a new entry is updated / created it
will be reflected here.

entries/xxxx-yyyy-zzzz.md files are the actual source content, the way
I'm doing it right now is very **DANGEROUS** and should never be used,
that is, to plain save the user input without sanatizing it, the
potential problem is someone can write content like a virus, or a
trojan. Think that someone can write a "blog" whose content turns into a
program that deletes the server or worse yet, steals information (see
https://www.owasp.org/index.php/Category:OWASP_Top_Ten_2017_Project )
specifically https://www.owasp.org/index.php/Top_10-2017_A1-Injection