Create html from input and displays it
This is a big push on the project.

The original index.html has been renamed to public/example-file.html and
now the generated entries are using the CSS defined earlier.

The user input is saved in two formats

- `/entries/file-name.json`
- `/generated/file-name.html`

The entries are also saved in `/entries/index.json` for to make it
easier to work with them.

A very basic "include" mechanism is added in this commint, is not meant
to be the final form, but it adds the common pieces to the entry: the
header and footer.

To list all the entries an iframe is used that converts
`/entries/index.json` into an html list.

Closes #6 #14 #8 #12 #10