Takes input from prosemirror editor
It turns out that the content of the basic editor already delivers HTML
that can then be used in the existing /api endpoint.

This change adds a temporal function in js/main.js "saveDocument" that
takes the innerHTML and innertText from the editable area and sends it
so the new .html file can be created.

At this point, the current window.document is completely replaced with
the response from the server, but this won't be the case in the future,
where a redirect to the correct Url will be more appropiate.

Also temporarly is the use of CSS in the prosemirror directly, this
might or might not be the final way to add styling, at this moment it
looks ok.

The title in the includes/header.htlm file is commented out for the time
being.

The old text area is completely deprecaed and was renated to
`input-old.html`