Encode/Decode content to avoid losing data.
While testing some blogs I noticed the content was getting lose due to
the existance of and ampersand inserted by the editor.

e.g. some spaces are represented as `&nbsp` and when the data was being
sent the content truncated there.

I also did an attempt to use display the title and rearranged a couple
of things while there.