Fixed the deployment.
The paths are to be relative instead of absolute for we don't know if we
run in a context (for instance oscarryz.github.io/manu) so don't assume
we can access the domain root path `/`

Also, when running completely serverless, there's not redirect from / to
the latest entry, so now when a new entry is updated, the
`generated/index.html` file is updated with an HTML redirect to this new
file.

A deploy script was ~improvised~ created to copy all the files from
`public` and `generated` into `dist` which then will be used by
`gh-pages` command.